requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Pillow==9.5.0
numpy==1.24.3
torch==2.0.0
torchvision==0.15.0